Move UsersTable to Admin controller
