+++ EPlast/EPlast/Views/Admin/Index.cshtml
@@ -7,15 +7,10 @@
         {
             <tr>
                 <td>
-                    <a class="context-menu-one btn btn-neutral "data-id="@user.Id">@user.Email</a>
+                    <a class="context-menu-one btn btn-neutral " data-id="@user.Id">@user.Email</a>
                 </td>
             </tr>
 
         }
-
     </table>
-
-    <div id="modDialog" class="modal fade">
-        <div id="dialogContent" class="modal-dialog"></div>
-    </div>
 </div>
