+++ EPlast/EPlast/Views/Admin/UsersTable.cshtml
@@ -4,7 +4,7 @@
     ViewData["Title"] = "Таблиця корустувачів";
 }
 <div class="container">
-    <div class="data-table-wrapper">
+    <div class="data-table-wrapper admin-edit">
         <table id="dtReadRaport" class="table table-striped table-bordered table-sm" cellspacing="0" width="100%">
             <thead>
                 <tr>
@@ -40,22 +40,27 @@
             <tbody>
                 @foreach (var item in Model)
                 {
-                <tr>
-                    <td>@item.User.LastName</td>
-                    <td>@item.User.FirstName</td>
-                    <td>@if (item.User.UserProfile.GenderID == 1)
-                    {<span>Чоловік</span>}
-                    else
-                    {<Span>Жінка</Span>}</td>
-                    <td>@item.UserPlastDegreeName</td>
-                    <td>@item.User.Email</td>
-                    <td>@item.RegionName</td>
-                    <td>@item.CityName</td>
-                    <td>@item.ClubName</td>
-                    <td>@item.UserRoles</td>
-                </tr>
+                    <tr class="context-menu-one" data-id="@item.User.Id">
+                        <td>@item.User.LastName</td>
+                        <td>@item.User.FirstName</td>
+                        <td>
+                            @if (item.User.UserProfile.GenderID == 1)
+                            {<span>Чоловік</span>}
+                            else
+                            {<Span>Жінка</Span>}
+                        </td>
+                        <td>@item.UserPlastDegreeName</td>
+                        <td>@item.User.Email</td>
+                        <td>@item.RegionName</td>
+                        <td>@item.CityName</td>
+                        <td>@item.ClubName</td>
+                        <td>@item.UserRoles</td>
+                    </tr>
                 }
             </tbody>
         </table>
+        <div id="modDialog" class="modal fade">
+            <div id="dialogContent" class="modal-dialog"></div>
+        </div>
     </div>
 </div>
